requirements.txt
alembic==1.8.0
bcrypt==3.2.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
firebase-admin==5.2.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-logging==3.2.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
oauth2client==4.1.3
postgres==4.0
proto-plus==1.22.0
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.37
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
